switch into midpoint_circle

+++ Include/Point.h
@@ -1,6 +1,7 @@
 #ifndef POINT_H
 #define POINT_H
 #include<cmath>
+
 struct Point
 {
 	int x;
@@ -9,9 +10,9 @@ struct Point
 	Point(int x, int y) { this->x = x; this->y = y; }
 };
 
-static inline double distance(Point &x0, Point &x1)
+static inline int distance(Point &x0, Point &x1)
 {
-   return sqrt ( (x1.x - x0.x) *  (x1.x - x0.x) + (x1.y - x0.y) *  (x1.y - x0.y));
+   return (int) sqrt ( (x1.x - x0.x) *  (x1.x - x0.x) + (x1.y - x0.y) *  (x1.y - x0.y));
 }
 
 #endif
